+++ packages/scenes/src/variables/variants/DataSourceVariable.test.ts
@@ -74,7 +74,7 @@ describe('DataSourceVariable', () => {
         options: [],
         value: '',
         text: '',
-        query: '',
+        pluginId: '',
       });
 
       await lastValueFrom(variable.validateAndUpdate());
@@ -92,7 +92,7 @@ describe('DataSourceVariable', () => {
         options: [],
         value: '',
         text: '',
-        query: 'non-existant-datasource',
+        pluginId: 'non-existant-datasource',
       });
 
       await lastValueFrom(variable.validateAndUpdate());
@@ -113,7 +113,7 @@ describe('DataSourceVariable', () => {
         options: [],
         value: '',
         text: '',
-        query: 'prometheus',
+        pluginId: 'prometheus',
       });
 
       await lastValueFrom(variable.validateAndUpdate());
@@ -142,7 +142,7 @@ describe('DataSourceVariable', () => {
         options: [],
         value: '',
         text: '',
-        query: 'prometheus',
+        pluginId: 'prometheus',
       });
 
       await lastValueFrom(variable.validateAndUpdate());
@@ -164,7 +164,7 @@ describe('DataSourceVariable', () => {
         options: [],
         value: '',
         text: '',
-        query: 'prometheus',
+        pluginId: 'prometheus',
         regex: 'slow.*',
       });
 
@@ -181,7 +181,7 @@ describe('DataSourceVariable', () => {
         options: [],
         value: '',
         text: '',
-        query: 'prometheus',
+        pluginId: 'prometheus',
         regex: '$variable-1.*',
       });
 
@@ -198,7 +198,7 @@ describe('DataSourceVariable', () => {
       const variable = new DataSourceVariable({
         name: 'test',
         options: [],
-        query: 'prometheus',
+        pluginId: 'prometheus',
         value: 'slow-prometheus-mocked',
         text: 'slow-prometheus-mocked',
       });
@@ -214,7 +214,7 @@ describe('DataSourceVariable', () => {
         name: 'test',
         options: [],
         isMulti: true,
-        query: 'prometheus',
+        pluginId: 'prometheus',
         value: ['prometheus-mocked', 'slow-prometheus-mocked', 'elastic-mocked'],
         text: ['prometheus-mocked', 'slow-prometheus-mocked', 'elastic-mocked'],
       });
@@ -230,7 +230,7 @@ describe('DataSourceVariable', () => {
         name: 'test',
         options: [],
         isMulti: true,
-        query: 'elastic',
+        pluginId: 'elastic',
         value: ['prometheus-mocked', 'slow-prometheus-mocked'],
         text: ['prometheus-mocked', 'slow-prometheus-mocked'],
       });